feat: payment setup

+++ apps/api/src/order/OrderApiService.ts
@@ -1,10 +1,16 @@
 import { Cart } from '@app/entity/domain/cart/Cart.entity';
 import { CartItem } from '@app/entity/domain/cart/CartItem.entity';
 import { Order } from '@app/entity/domain/order/Order.entity';
-import { OrderItem } from '@app/entity/domain/order/OrderItem.entity';
+import {
+  Payment,
+  PaymentDocument,
+} from '@app/entity/domain/payment/Payment.schema';
+import { PaymentService } from '@app/entity/domain/payment/PaymentService';
 import { User } from '@app/entity/domain/user/User.entity';
 import { Injectable } from '@nestjs/common';
+import { InjectModel } from '@nestjs/mongoose';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Model } from 'mongoose';
 import { In, Repository } from 'typeorm';
 
 @Injectable()
@@ -15,6 +21,9 @@ export class OrderApiService {
     private readonly cartItemRepository: Repository<CartItem>,
     @InjectRepository(Order)
     private readonly orderRepository: Repository<Order>,
+    @InjectModel(Payment.name)
+    private readonly paymentModel: Model<PaymentDocument>,
+    private readonly paymentService: PaymentService,
   ) {}
 
   private async findCartItemsByIds(cartItemIds: number[]): Promise<CartItem[]> {
@@ -26,25 +35,54 @@ export class OrderApiService {
     });
   }
 
-  async orderFromCart(user: User, cartItemIds: number[]): Promise<Order> {
-    const cart = await this.cartRepository.findOne({
+  private async findCartWithItemsByUser(user: User): Promise<Cart> {
+    return await this.cartRepository.findOne({
       where: {
         user,
       },
       relations: ['items'],
     });
+  }
 
-    if (!cart.hasCartItems(cartItemIds)) {
-      return null;
+  async orderFromCart(
+    user: User,
+    cartItemIds: number[],
+    impUid: string,
+    merchantUid: string,
+  ): Promise<Order> {
+    try {
+      const cart = await this.findCartWithItemsByUser(user);
+      if (!cart.hasCartItems(cartItemIds)) {
+        return null;
+      }
+
+      const { code, message, response } = await this.paymentService.complete(
+        impUid,
+      );
+
+      const cartItems = await this.findCartItemsByIds(cartItemIds); // lazy ?
+      const order = Order.create(merchantUid, user, cartItems);
+      if (response.amount !== order.getTotalAmount()) {
+        throw new Error();
+      }
+
+      switch (response.status) {
+        case 'paid': {
+          const payment = new this.paymentModel(response);
+          await payment.save();
+
+          // tx start ?
+          await this.orderRepository.save(order);
+          await this.cartItemRepository.remove(cartItems);
+          // tx end
+          return order;
+        }
+        default:
+          // 결제 취소? payment  삭제?
+          throw new Error();
+      }
+    } catch (err) {
+      // 결제 완료 에러
     }
-
-    const cartItems = await this.findCartItemsByIds(cartItemIds);
-    const order = await this.orderRepository.save(
-      Order.create(user, cartItems),
-    );
-
-    await this.cartItemRepository.remove(cartItems);
-
-    return order;
   }
 }
